fix: 카테고리별 스팟 API CategoryId 타입 Int로 수정

+++ data/src/main/java/com/cmc/data/feature/spot/repository/SpotRepositoryImpl.kt
@@ -19,7 +19,7 @@ class SpotRepositoryImpl @Inject constructor(
 ): SpotRepository {
 
     override suspend fun getCategorySpots(
-        categoryId: Long,
+        categoryId: Int,
         latitude: Double,
         longitude: Double,
         sortBy: String
